Split up incremental and tree-shaped tokenizers into separate classes
