+++ org.spoofax.jsglr2/src/main/java/org/spoofax/jsglr2/JSGLR2Variant.java
@@ -19,8 +19,8 @@ import org.spoofax.jsglr2.stack.collections.ActiveStacksRepresentation;
 import org.spoofax.jsglr2.stack.collections.ForActorStacksRepresentation;
 import org.spoofax.jsglr2.tokens.StubTokenizer;
 import org.spoofax.jsglr2.tokens.TokenizerVariant;
-import org.spoofax.jsglr2.tokens.TreeTokens;
-import org.spoofax.jsglr2.tokens.incremental.IncrementalTreeTokens;
+import org.spoofax.jsglr2.tokens.incremental.IncrementalTreeShapedTokenizer;
+import org.spoofax.jsglr2.tokens.treeshaped.TreeShapedTokenizer;
 
 public class JSGLR2Variant {
     public final ParserVariant parser;
@@ -55,9 +55,9 @@ public class JSGLR2Variant {
             case Iterative:
                 return new IterativeStrategoTermTokenizer();
             case TreeShaped:
-                return new TreeTokens.Tokenizer();
+                return new TreeShapedTokenizer();
             case IncrementalTreeShaped:
-                return new IncrementalTreeTokens.Tokenizer();
+                return new IncrementalTreeShapedTokenizer();
             default:
             case Null:
                 throw new IllegalStateException();
